+++ src/functions/private/Expand-FlatPermissionReference.ps1
@@ -7,8 +7,8 @@ function Expand-FlatPermissionReference {
         $SortedPath,
         $PrincipalsByResolvedID,
         $ACEsByGUID,
-        $AceGUIDsByPath,
-        [hashtable]$ShortNameByID = [hashtable]::Synchronized(@{})
+        $AceGUIDsByPath##,
+        ##[hashtable]$ShortNameByID = [hashtable]::Synchronized(@{})
 
     )
 
@@ -20,7 +20,7 @@ function Expand-FlatPermissionReference {
 
         ForEach ($ACE in $ACEsByGUID[$AceGUIDs]) {
 
-            Merge-AceAndPrincipal -ACE $ACE -Principal $PrincipalsByResolvedID[$ACE.IdentityReferenceResolved] -PrincipalByResolvedID $PrincipalsByResolvedID -ShortNameByID $ShortNameByID
+            Merge-AceAndPrincipal -ACE $ACE -Principal $PrincipalsByResolvedID[$ACE.IdentityReferenceResolved] -PrincipalByResolvedID $PrincipalsByResolvedID ##-ShortNameByID $ShortNameByID
 
         }