+++ src/functions/private/Group-ItemPermissionReference.ps1
@@ -5,9 +5,9 @@ function Group-ItemPermissionReference {
         $AceGUIDsByPath,
         $ACEsByGUID,
         $PrincipalsByResolvedID,
-        [hashtable]$Property = @{},
-        [hashtable]$IdByShortName = [hashtable]::Synchronized(@{}),
-        [hashtable]$ShortNameByID = [hashtable]::Synchronized(@{})
+        [hashtable]$Property = @{}##,
+        ##[hashtable]$IdByShortName = [hashtable]::Synchronized(@{}),
+        ##[hashtable]$ShortNameByID = [hashtable]::Synchronized(@{})
     )
 
     ForEach ($ItemPath in $SortedPath) {
@@ -15,10 +15,13 @@ function Group-ItemPermissionReference {
         $Property['Path'] = $ItemPath
         $IDsWithAccess = Find-ResolvedIDsWithAccess -ItemPath $ItemPath -AceGUIDsByPath $AceGUIDsByPath -ACEsByGUID $ACEsByGUID -PrincipalsByResolvedID $PrincipalsByResolvedID
 
-        $Property['Access'] = ForEach ($ShortName in ($ShortNameByID[$IDsWithAccess.Keys] | Sort-Object)) {
+        ##$Property['Access'] = ForEach ($ShortName in ($ShortNameByID[$IDsWithAccess.Keys] | Sort-Object)) {
+        $Property['Access'] = ForEach ($ID in ($IDsWithAccess.Keys | Sort-Object)) {
             [PSCustomObject]@{
-                Account  = $ShortName
-                AceGUIDs = $IDsWithAccess[$IdByShortName[$ShortName]]
+                ##Account  = $ShortName
+                ##AceGUIDs = $IDsWithAccess[$IdByShortName[$ShortName]]
+                Account  = $ID
+                AceGUIDs = $IDsWithAccess[$ID]
             }
         }