+++ src/functions/private/Group-TargetPermissionReference.ps1
@@ -14,10 +14,10 @@ function Group-TargetPermissionReference {
 
         # How to group the permissions in the output stream and within each exported file
         [ValidateSet('account', 'item', 'none', 'target')]
-        [string]$GroupBy = 'item',
+        [string]$GroupBy = 'item'##,
 
-        [hashtable]$IdByShortName = [hashtable]::Synchronized(@{}),
-        [hashtable]$ShortNameByID = [hashtable]::Synchronized(@{})
+        ##[hashtable]$IdByShortName = [hashtable]::Synchronized(@{}),
+        ##[hashtable]$ShortNameByID = [hashtable]::Synchronized(@{})
 
     )
 
@@ -68,20 +68,21 @@ function Group-TargetPermissionReference {
                     }
 
                     $AceGuidByIDForThisNetworkPath = @{}
-                    $IdByShortNameForThisTarget = @{}
+                    ##$IdByShortNameForThisTarget = @{}
 
                     ForEach ($ID in $IDsWithAccess.Keys) {
 
-                        $ShortName = $null
-                        $ShortName = $ShortNameByID[$ID]
+                        ##$ShortName = $null
+                        ##$ShortName = $ShortNameByID[$ID]
 
-                        if (-not $ShortName) { continue }
-                        $IdByShortNameForThisTarget[$ShortName] = $IdByShortName[$ShortName]
+                        ##if (-not $ShortName) { continue }
+                        ##$IdByShortNameForThisTarget[$ShortName] = $IdByShortName[$ShortName]
 
-                        $IdentityString = [string]$ID
+                        ##$IdentityString = [string]$ID
                         $GuidsForThisIDAndNetworkPath = [System.Collections.Generic.List[guid]]::new()
 
-                        ForEach ($Guid in $AceGUIDsByResolvedID[$IdentityString]) {
+                        ##ForEach ($Guid in $AceGUIDsByResolvedID[$IdentityString]) {
+                        ForEach ($Guid in $AceGUIDsByResolvedID[$ID]) {
 
                             $AceContainsThisID = $AceGuidsForThisNetworkPath[$Guid]
 
@@ -91,13 +92,14 @@ function Group-TargetPermissionReference {
 
                         }
 
-                        $AceGuidByIDForThisNetworkPath[$IdentityString] = $GuidsForThisIDAndNetworkPath
+                        $AceGuidByIDForThisNetworkPath[$ID] = $GuidsForThisIDAndNetworkPath
 
                     }
 
                     [PSCustomObject]@{
                         Path     = $NetworkPath
-                        Accounts = Group-AccountPermissionReference -ID $IdByShortNameForThisTarget -AceGuidByID $AceGuidByIDForThisNetworkPath -AceByGuid $ACEsByGUID
+                        #Accounts = Group-AccountPermissionReference -ID $IdByShortNameForThisTarget -AceGuidByID $AceGuidByIDForThisNetworkPath -AceByGuid $ACEsByGUID
+                        Accounts = Group-AccountPermissionReference -ID $ID -AceGuidByID $AceGuidByIDForThisNetworkPath -AceByGuid $ACEsByGUID
                     }
 
                 }
@@ -122,10 +124,12 @@ function Group-TargetPermissionReference {
                 $TargetProperties['NetworkPaths'] = ForEach ($NetworkPath in $NetworkPaths) {
 
                     $TopLevelItemProperties = @{
-                        'Items' = Group-ItemPermissionReference -SortedPath ($Children[$NetworkPath] | Sort-Object) -ACLsByPath $ACLsByPath -IdByShortName $IdByShortName -ShortNameByID $ShortNameByID @CommonParams
+                        ##'Items' = Group-ItemPermissionReference -SortedPath ($Children[$NetworkPath] | Sort-Object) -ACLsByPath $ACLsByPath -IdByShortName $IdByShortName -ShortNameByID $ShortNameByID @CommonParams
+                        'Items' = Group-ItemPermissionReference -SortedPath ($Children[$NetworkPath] | Sort-Object) -ACLsByPath $ACLsByPath @CommonParams
                     }
 
-                    Group-ItemPermissionReference -SortedPath $NetworkPath -Property $TopLevelItemProperties -ACLsByPath $ACLsByPath -IdByShortName $IdByShortName -ShortNameByID $ShortNameByID @CommonParams
+                    ##Group-ItemPermissionReference -SortedPath $NetworkPath -Property $TopLevelItemProperties -ACLsByPath $ACLsByPath -IdByShortName $IdByShortName -ShortNameByID $ShortNameByID @CommonParams
+                    Group-ItemPermissionReference -SortedPath $NetworkPath -Property $TopLevelItemProperties -ACLsByPath $ACLsByPath @CommonParams
 
                 }
 
@@ -155,7 +159,8 @@ function Group-TargetPermissionReference {
 
                     [PSCustomObject]@{
                         Path   = $NetworkPath
-                        Access = Expand-FlatPermissionReference -SortedPath $ItemsForThisNetworkPath -ShortNameByID $ShortNameByID @CommonParams
+                        Access = Expand-FlatPermissionReference -SortedPath $ItemsForThisNetworkPath @CommonParams
+                        ##Access = Expand-FlatPermissionReference -SortedPath $ItemsForThisNetworkPath -ShortNameByID $ShortNameByID @CommonParams
                     }
 
                 }