+++ src/functions/private/Merge-AceAndPrincipal.ps1
@@ -3,26 +3,27 @@ function Merge-AceAndPrincipal {
     param (
         $Principal,
         $ACE,
-        $PrincipalByResolvedID,
-        [hashtable]$ShortNameByID = [hashtable]::Synchronized(@{})
+        $PrincipalByResolvedID##,
+        ##[hashtable]$ShortNameByID = [hashtable]::Synchronized(@{})
     )
 
     ForEach ($Member in $Principal.Members) {
 
-        Merge-AceAndPrincipal -ACE $ACE -Principal $PrincipalByResolvedID[$Member] -PrincipalByResolvedID $PrincipalByResolvedID -ShortNameByID $ShortNameByID
+        Merge-AceAndPrincipal -ACE $ACE -Principal $PrincipalByResolvedID[$Member] -PrincipalByResolvedID $PrincipalByResolvedID ##-ShortNameByID $ShortNameByID
 
     }
 
-    $AccountName = $ShortNameByID[$Principal.ResolvedAccountName]
+    ##$AccountName = $ShortNameByID[$Principal.ResolvedAccountName]
 
-    # If the Export-Permission parameters excluded this principal it will not exist in the $ShortNameByID hashtable so we do not want to return this ACE; return nothing instead.
-    if (-not $AccountName) { return }
+    ## # If the Export-Permission parameters excluded this principal it will not exist in the $ShortNameByID hashtable so we do not want to return this ACE; return nothing instead.
+    ##if (-not $AccountName) { return }
 
     $OutputProperties = @{
         PSTypeName  = 'Permission.FlatPermission'
         ItemPath    = $ACE.Path
         AdsiPath    = $Principal.Path
-        AccountName = $AccountName
+        AccountName = $Principal.ResolvedAccountName
+        ##AccountName = $AccountName
     }
 
     ForEach ($Prop in ($ACE | Get-Member -View All -MemberType Property, NoteProperty).Name) {